Classroom-bot/pyproject.toml
[tool.poetry]
name = "repl_python3_Classroom-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
aiohttp = "^3.7.4"
lxml = "^4.6.3"
pytz = "^2021.1"
urbandictionary = "^1.1"
wikipedia = "^1.4.0"
python-dotenv = "^0.19.0"
google-search-results = "^2.4.0"
schedule = "^1.1.0"
discord-components = "^2.1.2"
"discord.py" = "^1.7.3"
"slash-discord.py" = "^1.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
